+++ packages/dms-builder/src/implementations/dms-builder/engine/query-template.ts
@@ -150,7 +150,7 @@ function findField(
   return fields.find((field) => field.name === name);
 }
 
-export interface FieldTypes {
+interface FieldTypes {
   ts: string;
   dataType: string;
   /**