+++ packages/dms-builder/src/index.ts
@@ -14,7 +14,11 @@ import {
   FRONTEND_MODULE_PRIORITY,
 } from "./constants/routes";
 
-export type { BuilderApiConfig, DmsBuilderConfig } from "./config";
+export type {
+  BuilderApiConfig,
+  DmsBuilderConfig,
+  FactoringConfig,
+} from "./config";
 
 export async function construct(config?: DmsBuilderConfig): Promise<void> {
   setModuleConfig(config);